Electronic Entertainment Expo Project
This is a class project based on one of the most popular video game events of the year known as the Electronic Entertainment Expo, or E3. This website provides an overview of the event and what to expect, as well as a contact page to get in touch with the event itself.